requirements.txt
# Backend
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.32
httpx==0.27.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
python-dateutil==2.9.0

# Broker & Market Data
requests==2.32.3
requests-oauthlib==2.0.0
yfinance==0.2.43
pandas-ta==0.3.14b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

